mein_projekt/pyproject.toml
[project]
name = "mein-projekt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "brother-ql>=0.9.4",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
]
